Move SUMOSuggest into separate module
This is somewhat cosmetic. This structure makes it easier to have more
complicated suggestion providers.
