+++ fjord/suggest/providers/sumo/__init__.py
@@ -1,0 +1,6 @@
+from fjord.suggest.providers.sumo.provider import (
+    SUMOSuggestRedirector,
+    SUMOSuggest
+)
+
+__all__ = ['SUMOSuggestRedirector', 'SUMOSuggest']
